Inline the function `with_lvalue_ref` into `trans_transmute`
The function was a footgun because it created `undef` references to
ZSTs, which could cause trouble were they to leak to user code.
